Move interfaces to inner class
