+++ app/src/main/java/be/ugent/zeus/hydra/common/main/MainActivity.java
@@ -564,4 +564,43 @@ public class MainActivity extends BaseActivity implements NavigationView.OnNavig
             this.menuItem = menuItem;
         }
     }
+
+    /**
+     * Allow fragments to extract arguments from the activity.
+     *
+     * @author Niko Strijbol
+     */
+    @FunctionalInterface
+    public interface ArgumentsReceiver {
+
+        /**
+         * Called when the fragment is created by the hosting activity. Allows the fragment to extract arguments from
+         * the {@code activityIntent} and put them in {@code existingArguments}. The resulting bundle will then
+         * eventually be set as the arguments of the fragment.
+         *
+         * This function should be a pure function, meaning there should be no side effects in the fragment. Side-effects
+         * resulting from this function may cause undefined behaviour.
+         *
+         * @param activityIntent The intent of the activity.
+         * @param existingArguments The bundle to put the arguments in.
+         */
+        void fillArguments(Intent activityIntent, Bundle existingArguments);
+    }
+
+    /**
+     * Used to notify fragments they will be removed. At this point, a fragment should consider itself removed,
+     * even if it is not yet removed.
+     *
+     * This allows fragments to do things like dismiss snackbars, or close action modes.
+     *
+     * @author Niko Strijbol
+     */
+    @FunctionalInterface
+    public interface ScheduledRemovalListener {
+
+        /**
+         * Called when the user will be switching to another fragment.
+         */
+        void onRemovalScheduled();
+    }
 }
