+++ app/src/main/java/be/ugent/zeus/hydra/minerva/announcement/unreadlist/UnreadAnnouncementsFragment.java
@@ -22,11 +22,11 @@ import be.ugent.zeus.hydra.common.arch.observers.ErrorObserver;
 import be.ugent.zeus.hydra.common.arch.observers.ProgressObserver;
 import be.ugent.zeus.hydra.common.arch.observers.SuccessObserver;
 import be.ugent.zeus.hydra.common.database.RepositoryFactory;
+import be.ugent.zeus.hydra.common.main.MainActivity;
 import be.ugent.zeus.hydra.common.ui.BaseActivity;
 import be.ugent.zeus.hydra.common.ui.recyclerview.EmptyViewObserver;
 import be.ugent.zeus.hydra.common.ui.recyclerview.ResultStarter;
 import be.ugent.zeus.hydra.common.ui.recyclerview.adapters.MultiSelectDiffAdapter;
-import be.ugent.zeus.hydra.common.main.ScheduledRemovalListener;
 import be.ugent.zeus.hydra.minerva.announcement.Announcement;
 import be.ugent.zeus.hydra.minerva.announcement.SingleAnnouncementActivity;
 import be.ugent.zeus.hydra.minerva.announcement.courselist.AnnouncementsForCourseFragment;
@@ -47,7 +47,7 @@ import static android.app.Activity.RESULT_OK;
  *
  * @author Niko Strijbol
  */
-public class UnreadAnnouncementsFragment extends Fragment implements MultiSelectDiffAdapter.Callback<Announcement>, ScheduledRemovalListener {
+public class UnreadAnnouncementsFragment extends Fragment implements MultiSelectDiffAdapter.Callback<Announcement>, MainActivity.ScheduledRemovalListener {
 
     private static final String TAG = "UnreadAnnouncementsFragment";
